django/requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
duckdb==0.10.0
numpy==1.26.4
psycopg2==2.9.9
pyarrow==15.0.0
pytz==2024.1
sqlparse==0.4.4
